+++ packages/map-template/src/components/WebComponentWrappers/Notification/Notification.scss
